Refactored models to use BigDecimal for precision.

+++ src/main/java/com/shopi/shopping/models/products/Book.java
@@ -2,12 +2,15 @@ package com.shopi.shopping.models.products;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
+
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "books")  // Maps this class to the "books" table in the database
 public class Book extends Product {
 
     // Constructor to initialize the book product
-    public Book(double price, String name, String description) {
+    public Book(BigDecimal price, String name, String description) {
         super(price, name, description);
     }
